% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{subset_incorrect_or_wrong_spot}
\alias{subset_incorrect_or_wrong_spot}
\title{Subset words from input}
\usage{
subset_incorrect_or_wrong_spot(wrong_spot, incorrect_letters)
}
\description{
Subset words from input
}
